requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-dotenv>=1.0.0
huggingface-hub>=0.20.0
